Refactor DataPreprocessor
Split out data discovery and validation so we can keep the pipeline
decoupled. That way the same discovery algorithm can be used in
different pipelines, just like the same validation can be used with a
different discovery (i.e as long as the same json input can be given to
the validator, even if the directory structure is different, we can
validate the input dataset the same way.

Currently the expectation is as follows:
`--root_dir`
```
some.shp
some_related.dbf,prj etc
dir1/
  site1/
    irrelevant_dir1/
    bamboo(irrelevant_dir2)/
    site1.tiff --> this needs to intersect the polygons in some.shp
  site2/
    site2.tiff --> this needs to intersect the polygons in some.shp
  site3/...
```
* The discovery class with create a json that has the shp file and the
site.tiff files.
* The validation class will take json that has a pointer the the main
  shp file as well as a list of tiff files, and ensure that the bounds
of each tiff intersect with polygons in the shp file
* The validation class also returns the total classes:counts in these
  intersecting polygons

There is no data transformation. All the validation script does is make
sure the data in the tiffs matches the data in the shp, for if this is
not true, all subsequent stages will fail.