Raise ValidationExceptions
1. Pass the name_key through discovery and validation. The name_key is
   taken from the user and represents the name under which the class
labels are stored in the shp file.

2. Raise a ValidationException if there is no intersection between the
   shp file and any of the site tiff files. This indicates a mismatch in
CRS or that the data set is corrupted in some way.

Why raise validation errors? rather than train the model with a
corrupted dataset, we want to stop and draw attention to the mismatch.

Opened https://github.com/bprashanth/dlt/issues/8 to deal with partial
mismatch of tiff x shp files.