Add a tile generator
* Tiling happens in one go using rasterio.
* The output of the discovery stage is used to tile.
* The shp file is ignored, but the tiff files are split into tiles.

Each tiff file has has the metadata needed to perform the tiling, i.e.

1. The CRS (e.g. `EPSG:32643`)
2. The transform to convert pixels to the CRS (e.g. affine)

The CRS is stored per image because it's a lookup into a database
`EPSG:32643` -> "WGS 84 / UTM zone 43N, in meters"
(https://epsg.io/32643) which says that the value in the tile_bounds is
a UTM projection in zone 43N on the datum WGS84.

At the end of tiling, we get a tiling directory which is laid out as
follows
```
output/
├── images/
│   ├── Gavihalla_x0_y0.png
│   ├── Gavihalla_x384_y0.png
│   └── ...
└── tiles_metadata.json
```
Where tiles_metadata.json has the "origin" of the tile, both in pixels
and in geospatial coordinates e.g.
```
[
{
    "filename": "Gavihalla_x0_y0.png",
    "site": "Gavihalla",
    "tile_bounds": [x1, y1, x2, y2],
    "pixel_origin": [x, y],
    "crs": "EPSG:32643"
},
...
]
```