Skip mostly empty/empty tiles
Previously we were sliding the tiling window across the entire
orthomosaic. This had the undesired effect of producing many completely
empty/mostly empty tiles, which would have confused the training phase
and potentially resulted in labelling bias (the model would have started
looking for empties and treating transparent edges as an indicator of
lantana growth).

In order to avoid this, we:
1. Figure out what the "nodata" value is from the dataset (rasterio
   exposes a `read_mask` function that returns this value)
2. This is a mask of the entire source. This mask shows 1 for "contains
   pixel" and 0 for "nodata". In some cases it might record
"transparent" for nodata, but this is irrelevant.
3. Clip this mask to the bounds of the current sliding window.
4. Compute the % of nodata vs data pixels in the current window and
   compare against `--skip_threshold`.
5. If the tile is eg less than 10% real pixels (which is the default
   value of skip_threshold), discard it.

To revert this behavior, set `--skip_threshold` to `0`, which means even 0
valid data pixels will not result in skipping the tile. To only skip
competely empty tiles, set it to a small value like `0.001`.