Trainer + coco validations
* Trainer that reads coco annotations
* COCO validation class to check for inconsistencies
* A script to validate standalone coco annotations
* Bug fixes to `annotation_builder`
	- remove triple nested lists
	- fix index offsets

The main issues sorted by this pr (in addition to the actual trainer)
are as follows:
1. There were multiple index offsets being applied across the board in
   preperation for the training step. These are made consistent across
the board (see docs/categories.md)
2. The `annotation_builder` was generating triple nested polygon lists
   because it was confusing multi and single polygons. This was
resulting in corrupt coco annotations.
3. Due to a lack of validation this was manifesting as a model training
   failure. So a `CocoValidation` class is added, as well as a script to
trigger _only_ the validation of the annotations in
`hack/validate_coco.py`.
4. The docker setup to run the trainer: The trainer runs in a docker for
   a few different reasons
	a. It makes it easier to split into a webservice. I expect this
pipeline to be multi staged, with the first stage running on local and
the second running on a remote cloud host.
	b. It means the user doesn't have to install heavy ml libraries
locally, or deal with ml library mismatches.

However this means that the main.py that runs the pipeline needs to
invoke the container via `docker-py`.