Fix overlap as % of tile size
The tile size and overlap were defaulting to pixels. Rather than ask the
user to compute the % overlap required, just take the %overlap as a
percent and do it in code.

Unfortunately this means we previously forgot to specify overlap and
ended up training on 2048 tiles with 128 pixel overlap, which is far too
small to really capture any entire lantan bushes.