Generate "invalid" tiles
Previously if the % of image in a cropped tile was too low w would skip
it. However this creates problems for the final restitching back into a
mosaic. So this pr introduces a new field in `tile_metadata.json` called
`is_valid`. Tiles that are deemed invalid are still rendered as pngs in
the tiling directory, but recorded as `is_valid: false` in
`tile_metadata.json`.

The annotation_builder also accounts for these by skipping `is_valid:
false` tiles in the generation of the coco annotations that go into
training.

Also adds a comment on the old default value for `--overlap` so it can
be used to regenerate consistent tiles as needed.