Adds a tile stitcher
Tile stitcher re-stitiches tiles to form a complete map.
Also regenerates all tiles with the following modifications

1. Overlap set correctly to 25%
2. Invalid tiles are included in the regeneration

This means the annotations were regenerated, and while the model should
still be able to predict the outputs it might have to be retrained on
this new 25% overlap set.

This change also modifies main.py and run_inference.sh to account for
tile stitching.