Downscale images
The stitches orthomosaics were too big, so this pr adds a scaling ratio
argument and pipes it through the stitching stage to be used in
downscaling.

It also adds 2 flags, one to scale tiles and one to scale maps, because
scaling tiles by the same map factor removes too many pixels from it.